backend/orchestrator/requirements.txt
Flask==3.1.1
Flask-JWT-Extended==4.7.1
flask-smorest==0.46.1
Flask-SQLAlchemy==3.1.1
marshmallow==4.0.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
Werkzeug==3.1.3
Flask-APScheduler==1.13.1
requests==2.32.4
gunicorn==23.0.0
cyclonedx-python-lib==11.0.0
xmltodict==0.14.2
flask-cors==6.0.1